mediawiki.router: Remove publicly exposed oojs-router module
The wrapper only adds a single line (`exports = new OO.Router`),
which isn't worth a separate module. Consolidate the internal module
into the entry point and remove it.

In Gerrit, as per CodeSearch Everywhere [1], there is only one call,
in MultimediaViewer's mmv.js, fixed in Ice09b523dfb.

The mention in NearbyPages is a false positive relating to the npm
module used for local development.

No known on-wiki usage [2].

[1] https://codesearch.wmcloud.org/search/?q=oojs-router%5B%27%22%5D&excludeFiles=package.%2Ajson%7Cwikipedia.org%2Fspeed-tests
[2] https://global-search.toolforge.org/?q=oojs-router&regex=1&namespaces=2%2C4%2C8

Depends-On: Ice09b523dfbefd67493175554eeb525771e8f29d
Change-Id: I55ef3a659a515fb074fd44beed754ae2206a3499