SrConverter: add test case for initial and final roman numerals
A naive reading of the SrConverter code might lead one to believe that
final roman numerals aren't properly handled -- but split actually
returns an empty string in the trailing position in that case, so the
code works correctly.  Add a test case demonstrating that conversion
is handled correctly for initial and final roman numerals.

Change-Id: Ia4857cab74bbb581baf46a51a080cf098cbb61d8