mail: Move SpecialEmailUser business logic to a new class
This patch moves all the static methods in SpecialEmailUser responsible
for actually sending the email to a new class. The goal of this patch is
to *only* move the code without any other changes, either to the new or
old code, so that said changes will be easier to review when made all in
the same file, in another patch.

The ultimate goal is to make EmailUser a command class (like BlockUser
and the like) that uses DI and non-static methods, and that other code
that needs to send emails can use without interfacing with the special
page directly.

Bug: T265541
Change-Id: Ia18bd04198539da25ab7b42aa223868bd5cbe3d5