jquery.makeCollapsible: support toggle placeholder
Previously, makeCollapsible users could either accept the default
placement of the toggle button, or supply their own button anywhere in
the DOM and be fully responsible for it (contents, events, etc.). This
change adds a third option: specify the position in the DOM using a
placeholder element, but still let jQuery.makeCollapsible populate the
button to look just like a default button.

Bug: T329167
Change-Id: Ic2c222f139e56b4bd8c12490646d213fd9d25337