ParserTestRunner: don't silently skip missing files in --updateKnownFailures
Two bugs here: first, we were silently skipping a needed file update in
::updateKnownFailures() if the file didn't previously exist.  We're going
to still avoid writing the file if it didn't previously exist, out of an
abundance of caution, but at least we'll now fail noisily so the problem
can be fixed.

Second, the `--parsoid` flag was overriding the result of
::getFileSkipMessage() so we were processing files for
`--updateKnownFailures` that should have been skipped because they
are not marked parsoid-compatible.  This override made sense when
we were still debugging the integrated-mode parsoid support in the
ParserTestRunner, but it is not needed anymore.

Change-Id: Iba961ea327e54bb6bdc87399dbcba87cd57b6b20