OutputPage: Restore view-source access to former wfReportTime data
Follows-up Id069f913034 (11e0fba3b0), which merged two mw.config
script tags into one. This is great, but had as side-effect that
it placed the hostname/backendtime report at the end of a 1000+ char
long time, thus making it tedious if not impossible to access by
vertical scroll on view-source.

Restore the order of the merge by setting the debug vars first.

Change-Id: I91d240630dd8479f18b207c2a551812bb32755e6