Fix total breakage of wgCanonicalServer fallback
wgCanonicalServer is false by default. Setup.php tried to set it to the
result of a wfExpandUrl() call, which initialised the UrlUtils
service with the current value of wgCanonicalServer, i.e. false.
UrlUtils had a fallback for null but not false. So subsequent calls to
wfExpandUrl() would always use the empty string for the canonical
server. Probably since 472a914c63baaff93.

So, have UrlUtils handle the fallback, and have Setup.php ask UrlUtils
for the final value.

Bug: T325529
Change-Id: I99329e75bd53049ae84a8ae7a2536ad12cfeeb3b