rdbms: Add missing QUERY_CHANGE_ flag to internal "USE" query
The added flag makes no functional difference. It informs
executeQuery() and SQLPlatform->isWriteQuery() that this is not a
write query, which it would otherwise find out by running a regex
on the SQL text.

QUERY_IGNORE_DBO_TRX is documented as being for USE queries,
which in isWriteQuery() return false, just as it did already
based on the regex previously.

This follows-up to Ic91412ffdaab6b7d which made isWriteQuery warn
when this flag is missing, precisely to find this inefficiency.

Bug: T332228
Change-Id: If510415dbcf407353612f7b8e425d8a8189e26d9
(cherry picked from commit 8ce966af65bab3977eee8e1bbe18c2fdc62ae7c7)