import: Check for seekable streams before doing syntax check
The syntax check is done in a first read through the stream and
after a valid document is detected the stream is rewind to read the real
import from the begin.

This is not possible with php stdin as used by importDump.php without a
file as argument. As this is a maintenance scripts, it shoud be okay to
trust the source of xml and assume it is always correct syntax.
To check syntax it is necessary to use provide the filename to the
script and not pipe it into stdin.

Bug: T331992
Change-Id: I056f325bb2ccb9a0fe8245194baf2e73be13da68
(cherry picked from commit cac48ae1dacec09fb59c0e57708e201dc28364ea)