TempFSFile: Keep the WeakMap alive
If you store a WeakMap in the object, containing only circular
references in the values, then nothing is keeping the WeakMap alive and
it can be deleted along with its parent TempFSFile by the garbage
collector.

A WeakMap contains normal references to its values, so those normal
references can be marked by the GC like any other references.

Instead, store the WeakMap as a static property.

Bug: T332461
Bug: T332397
Change-Id: Idbaa1019c2bac3d1dd9b001753c2f4c0aaf66e98