objectcache: fix connections in RedisBagOStuff multi methods
Previously, a bug resulted in a connection per key for getMulti()
and similar methods. RedisConnRef instances are treated as if they
belong to different callers, so RedisConnectionPool::getConnection
cannot be used in naive loops to build (key => RedisConnRef) arrays.

Create a RedisBagOStuff::getConnectionsForKeys() method to properly
get connection handles for multiple keys and reduce code duplication.

Bug: T332371
Change-Id: Id58b9ac61c59277f7cd8f17821c1905b930b02cb