SpecialMyLanguage: Fix base page being ignored from language fallback detection
Fix the detection for language pages when the page language of the base page
is in the user interface language's fallback chain.

For example, if the user interface language is zh-hk and the base page is zh
and there are no zh-hk subpages / translations, it should redirect the user
to zh instead of en.

Bug: T333187
Change-Id: Ie425f13e1b361a1d05f90a1af074af5f4fc7dc6e