Add classes on elements inside the media structure
The purpose of which is to improve the performance of the css selectors
targeting the output, as analyzed in T270150#8524965, as well as
eliminate some of the brittleness in depending on direct descendents
and first-child, which can be seen in T320285 and T304010.

mw-file-element is targeted to apply margins, borders, and
vertical-alignment to that element.  The current css rules have wildcard
selectors in the rightmost position, which, since css is parsed
right-to-left, can be quite slow on a wiki page.  The legacy parser has
an equivalent class, thumbimage, when rendering thumbs but here we apply
the class more broadly.

A follow-up patch in I70c61493fe492445702f036e5b24ef87fc3bdf43 will
remove the redundant wildcard selectors once parsercache has turned
over.

Bug: T270150
Bug: T314097
Depends-On: Ie85ee7048273023a2c51f42a333a9c1493360404
Depends-On: Ie0ec018ac6c2c42c05610b342d7ef87493dfdc42
Depends-On: Ifc17fdf530af515b066de706ca5e69e118fd1c5b
Depends-On: Ib60edacdae2ff41a0de2b2b584718fd9ce925f97
Change-Id: Ifd4001e312a5fa4b7beaad63ba8c4e79e3201b9b