language: Annotate list() methods as preserving taintedness
The $list parameter to the various *list() methods is not escaped by
these methods, so any taintedness in the argument is preserved by the
method and kept in the return value. taint-check has troubles figuring
this out on its own due to T274780, so annotate the parameters.

Note: once this is merged, taint-check could start failing on some
repos. Any errors would have a decent chance of being true positives.

Bug: T253879
Change-Id: I6cf56aca9760370cbeae19879e6b170b1cbd273f