Add MessagesXx.php files for some RTL languages
These languages have some content in the Wikimedia Incubator
and/or Translatewiki, but because they don't have
MessagesXx.php files (with $rtl = true;), they get the
default LTR direction, and we needed to add some CSS &
Javascript hacks to get them to display correctly (and
even then, only partially).

The presence of these files does not enable these languages
in the preferences, it only sets the RTL setting for pages
with any of these languages as a page language.

Change-Id: I5c78889474d02850198519de04d2a553f1f97a9e