Parsoid PageConfigFactory: warn on any attempt to use deprecated argument
The 4th argument to PageConfigFactory::create() used to be $textOverride.
That argument now does nothing, but it may still entrap unwary users who
try to pass a text override to the page config factory.  It is hard to
remove this argument entirely due to the dependencies between core and
Parsoid, but we can at least wfDeprecated() anything which uses it.

Change-Id: I1cc4d2075016e6bd479cef2e32cb768ff1ad51df