qunit: Move <div id="qunit-fixture"> from JS to HTML
This is the recommendation upstream (admittedly edited by me, but
it's been that way for 10+ years) as per
<https://qunitjs.com/intro/#in-the-browser>.

The reason we did it in JS is because of karma-runner, which loads
the test suite directly into an empty browser tab without the visual
HTML page. Thus to get the fixture there, we had to create it
ourselves. But, this is a bug in the karma-qunit adapater as it is
meant to take responsibility for creating this.

And, since at least 2015, it does precisely that per
<https://github.com/karma-runner/karma-qunit/pull/22> so this is no
longer needed!

Bug: T250045
Change-Id: I2a2bf7d231d34bf54dc6c0ff3173eada74ae3ff1