Skin: Include dir for language links that do not match interface language
When listing language links, we should include the dir attribute
for languages which do not match the direction. This makes the page
more machine-readable and allows for different stylistic treatment if
desired.

This will be helpful
for the mobile language overlay which needs to currently use the API
to look up language direction.
Change-Id: I54376d8314a75d115a11425db9c5b8696d9515d7