ResourceLoader: Move ForeignResourceStructureTest to integration/
Follows-up I071be5821711f (2cf45bc934) and I8aeffc1f2e81 (3270283abf).

Avoid use of `@standalone` in core, since that group is only
special-cased in WMF CI for extension and skin repos.

Core tests, unlike extension tests, are already skipped by default
when running PHPUnit in CI for extension and skin repos. The exception
is the structure/ directory which is for general system-wide checks
that may also cover installed skins and extentensions.

Given that this test only covers core, and that's why we want to
skip it, move it out of structure/.

Bug: T203694
Change-Id: I7431e5627861845f9d49afc4777b76052d3a6e77