ResourceLoader: Set eslint to require ES3 syntax in startup module
This is very unique to only the startup module, no other module
needs to worry about this. As per the startup.js header, this module
should parse in ES3.

Since the 2018 consolidation of startup.js and mw.loader.js into
one module, this means it applies to the directory not just the
file. Although the rest of mw.loader can safely use any ES5 methods,
just not ES5 syntax.

Change-Id: Ib91f1b8028b97b351b9cfc14decf1ef1ce7106f9