ResourceLoader: Tweak startup.js known browsers explanation
Rephrase this so as to reduce confusion about what is and isn't
supported. First and foremost these are primarly a reflection of
objective reality with regards to passing the checks, not an intent
of support. These are either supported Grade A, or browsers that are
equivalent enough and work under Grade X (e.g. less well-known
distros of the same engines, such as UC Browser, Vivaldi, Iceweasel;
or beta versions, and older versions that are new enough for our
purposes). They will always pass the checks whether we want them to
or not.

In extreme cases, we may improve unofficial support for a Grade X
browser by moving it explicitly into Grade C through a UA sniff, but
as of 6c7177349bd there are no longer such cases, and if we keep up
the pace [1], then it seems likely that there won't be new entries
or that UA sniffs are short-lived, staying only until the next bump.

I've removed the Grade C listing from here as it no longer has a
definition or purpose. Before WMF wikis required HTTPS in 2015, this
was effectively a duplicate of the wiki page with some added entries
for the browsers we UA-sniffed against. Since then, many older
browsers have been cut off from Basic support for reasons unrelated
to CSS or JS, and so it has largely gone out of date, presumably
on the idea that startup.js should list a superset of the "official"
page, just like we do for the Grade A version, except it's not clear
what a superset of Grade C support means, but whatever it is, its
bottom cut-off is not related by the cut-off for the JS pipeline,
rather it's related to purely in-writing what we want to support.
There is an infinite list of text browsers and unsupported browsers
that would in theory get Grade C if they were able to connect by
other means, but that's neither here nor there.

[1] https://wikitech.wikimedia.org/wiki/Performance/Guides/Frontend_performance_practices#General_approach

Bug: T178356
Change-Id: I8e403123205fae2e7b15b03b8c6c0cc8e9013ea9