mediawiki.less: Deprecate Codex design tokens replaced variables
As mediawiki.skin.variables is featuring a number of equal keys,
let's deprecate the ones that are replaced by the new design tokens.
Note, that these are only deprecating same key and value and different
key, same value ones.
We're not removing the same key, same value ones right now, although
they technically could be removed as we plan to advertise around the
new mediawiki.skin.variables architecture first and stabilize it more.

Bug: T332541
Change-Id: Iddf76e8d428aac55df624c8ccf31dbf4e29b1ae4