Add showPendingTakeover method to OutputPage
This is upstreaming a pattern that's used in VisualEditor,
DiscussionTools, and (prompting this patch) MobileFrontend. These
extensions all take over certain views and want to avoid a flash of
non-taken-over content by using a hook to replace them with a "please
wait" loading message that'll redirect to a fallback page if the user
turns out to not have JavaScript enabled(or if it is but ResourceLoader
isn't available).

Change-Id: I8f25f53a05ce0437b0e9f63033cc151213c797bf