RELEASE-NOTES-1.40: Update and clean-up ahead of rc.0 branch
Most of this is moving content around to be in the proper place,
plus the occasional re-writing of entries to be more consistent
with our standard practices for these notes.

I've upgraded our claim for language support; Names::$names has 509
entries, and even when you take out the 80 sub-languages (en-AU,
zh-Hans, etc.) we're still way higher than 350 which we've had for
years now, so let's be more accurate.

Bug: T329082
Change-Id: I4b6611ed0b9f016e880b3db08e7d85e277ff209c