ResourceLoader: Remove 1/2 inDebugMode() calls in SpecialJavaScriptTest
This will be obsolete once we adopt version=2 but WMF CI isn't ready
for that yet per <https://phabricator.wikimedia.org/T250045#8671019>.

Meanwhile, we can reduce use of the global static method in one place.

SpecialJavaScriptTest has two main methods: renderPage() and exportJS().

renderPage() renders the HTML special page, and is where one might want
$wgResourceLoaderDebug to apply (which is the only thing that
inDebugMode does, in addition to just casting $_GET from string to
int, and then back to a string again, to be passed to RL\Context
to become an int again). This makes sense as OutputPage also uses
inDebugMode() like this.

exportJS() responds to the internal JS request made from the above
HTML page. This doesn't need to call this as it already gets the
debug= parameter set by the HTML. While impossible in practice, it
would make no sense for this to be in debug mode when the page isn't.
Like for index.php/OutputPage and load.php, we only call inDebugMode
on index.php. On load.php it gets debug= explicitly passed by the first
one and doesn't need to evaluate its own global state.

Bug: T250045
Bug: T36738
Change-Id: I2064a6cdc5970aec42d6bfaf70ea0c577642d465