rdbms: move DatabaseFactoryTest to unit/ directory
This was the last remaining test file for libs/rdbms/ outside
either phpunit/integration or phpunit/unit.

Move it to where it belongs to reduce the number of places to look
and avoid mistakes where running a directory appears to result in
meaningful tests but actually runs just this one test.

Change-Id: I8cada26b51bc191dc9b7be4dad7c8abbe107629e