rdbms: Make buildComparison() reject non-numeric indices
If there’s a numeric index, it’s probably a typo where someone put a
comma instead of an arrow; let’s make that an error instead of returning
some nonsensical SQL text that will do who knows what in the database.
(In the gate-and-submit for I10f91f7b09, patch set 1, such a mistake
produced a syntax error in the PostgreSQL build, but not in the MySQL or
SQLite ones, which is somewhat worrying.)

Also add tests for the other InvalidArgumentExceptions while we’re here.

Change-Id: Ia81c21b97d8406cee7c493b9fb6578a95cb39dcc