mediawiki.cookie: Factor out internal jar interface from $.cookie
Update these libraries so they do not depend on each other,
but instead create 2 entry points to the same library.

Follow up to I8cab36357535b3994e65f9ee81779479a72805d7

jquery.cookie was only using $.extend and given we are using
ES6 now, we can remove that dependency so this doesn't rely
on jQuery at all.

The library is renamed to jar.js and the header is updated to reflect
the fact that it has been forked from the upstream library.

References to $.cookie inside index.js now directly access
jar, however $.cookie is retained as a convenience function to access
the library to keep support with gadgets.

Bug: T271995
Change-Id: I47ad1c56d514502c468947ed7c22915740915f85