Make all pt/pt-br user namespace permutations aliases
Make "Usuária" (the pt-br version of the female user namespace)
an alias for the user namespace in `pt`, and make "Utilizador"
and "Utilizadora" aliases of the user namespace in `pt-br`.
Also add "Usuário(a)" and (by analogy) "Utilizador(a)" as
aliases; the former has been used as the namespace name in
ptwiki since around 2011 (it is difficult to trace down the
exact time), so it should be present in the default list
too.

This ensures that any link to a gendered user page on a wiki
set in either of the varieties will still end up at the right
place.

This patch is almost an exact duplication of patch 29761, which
was abandoned in 2014.

Bug: T33986
Change-Id: I539e781d44444994c2caf50d4a9eac8f13b0380b