ForeignResourceManager: Add transform step
Add a transformation step to foreign-resources.yaml that allows
for simple modifications after the files are fetched and hashed.

This is experimental and mainly meant as a short-term solution
for T335326. In the long term, we might decide to handle the
issue of libraries with ES2017 syntax differently, and this
functionality might be removed. Use at your own risk.

Currently the only supported modification is adding a 'nomin'
filter to the beginning of the file, which is useful for files
which are already minified or which the MediaWiki minifier can't
deal with for some reason. In the future, this could be used
e.g. to apply patches.

Change-Id: I4d93bf6d04683cc4072e99f6ab02d0e618d85914