build: Avoid extra space in Chromium command via CHROMIUM_FLAGS
When CHROMIUM_FLAGS is unset, as it is if you're running Chromium
outside any container (like fresh-node or Quibble), the splitting of
the empty string results not in an empty array but in an array with
an empty string. In other words `[ '' ]` insteadf of `[]`.

This means when Karma formats the browser launch command, and joins it
with a space, it runs `chromium --foo --bar   http://localhost`
notice the three spaces between "bar" and "http".

This was fine in Chromium for ten years but Chromium 110 changes the
way it parsers the command-line arguments such that it will interpret
this as launching Chromium with two URLs, of which one is the empty
space. While regular Chromium supports starting the browser with
multiple tabs at once, HeadlessChrome does not.

Bug: T335329
Change-Id: Ie8a1ee7a0b16ca6dfd7a8293e0bf686a48141c8c