Pass whether current rev of file exists to Linker::makeBrokenImageLinkObj
A $time parameter was added to Linker::makeBrokenImageLinkObj in 2d22c85
to avoid checking if a file exists if we were fetching the current
revision.  The assumption being that the method would only be called if
the current revision doesn't exist.  A new parameter is added to
indicate that the current revision exist so that a known link can be
created.

Bug: T329214
Change-Id: I9ae5b18c422ab7d1b2b3b599a1994c319fe37239