Fix frame and frameless rdfa depending on file existing
The mw:Error annotation is not specific to file handling.  For example,
it's set in the Cite extension.

Having the correct RDFa annotation is important in Parsoid (which the
fix here is matching) for roundtripping the format option.  It's also
targeted by the css selectors so that, even if the media fails to load,
the page layout maintains the editor's intention.

Change-Id: I669073cac0539c5e96bef1ea46156cc2ce730af9