upload: Improve report of invalid destination title on Special:Upload
Call the api with the File: namespace prefix to ensure that reports for
interwiki are correct. Everything starts with colon resets the namespace
in mw.Title to NS_MAIN and the title acts like main namespace and that
makes the interwiki prefix invalid. Giving this string to the api the
same happens on the server and api does not treat interwikis as invalid.
With File: the server knows what validation are needed.

Bug: T185904
Change-Id: I7f2eded599d0cfa1ab2424e771240026c97f1782