specials: Reject temp users on Special:Renameuser
Temp users should not be renamed to make them named like permanent users
When the new name is in the pattern for temp users it could be
problematic, when the next number is already there by a renamed account.

To rename temp users from the server side the maintenance script
renameUsersMatchingPattern.php exists.

Bug: T326926
Change-Id: I963c6b73253ed5cecd06c357ae8dd9a813e47552