rdbms: Throw when makeWhereFrom2d() receives empty data
This semi-internal method has only four callers, each of which
explicitly or implicitly ensures that $data is empty. For example:

* ApiQueryBacklinksprop: The caller returns early return if $titles
  is empty, and then proceeds to build $map from $titles PageSet.
  I've made it more explicit here for added confidence and to satisfy
  static analysis.
* cleanupPreferences.php: The caller explicilty checks $data
  (as $deleteWhere) before calling makeWhereFrom2d().
* ResourceLoader\WikiModule, the caller checks $batch->isEmpty()
  before calling $batch->constructSet().

More importantly, is that not one of these checks whether
makeWhereFrom2d() returns false, and blindly passes it on to
methods like IDatabase::select() or IDatabase::delete(), which does
not permit boolean false as $where/$conds, but also doesn't check
for it, which seems seems likely to wreak havoc at some point through
casting to empty string or empty array at some point.

Given there are very few callers, and given that the current
responsiblity to check the return value is never excercised, throw
an exception instead to make this code appear less fragile.

Follow-up CR from
<https://gerrit.wikimedia.org/r/c/mediawiki/core/+/892988>

Change-Id: Ifdf3ea54271ea856f2f555a5e087b8f521e348b3