editwatchlist.js: Use jQuery for performant select-all
Currently, the per-namespace check-all feature on Special:EditWatchlist
uses OOUI's selectItems. However, the performance is noticeably bad
with as few as 200 items, and hangs the browser with as few as 500.

This reimplements that step in jQuery, which performs smoothly with at
least 1000 items and should perform fine past the point where
Special:EditWatchlist runs into database limitations (T41510).

Bug: T335082
Bug: T334252
Change-Id: I5a9865f61ad08a18c1a3c6b73ea1837f657e0105