ResourceLoader: Log when MAXAGE_RECOVER is detected
Send a message to Logstash when this condition is reached.
It is expected that this happens for a portion of clients during the
5min after a deployment, such such use an informational message,
not a warning.

This level is already enabled in wmf-config for channel=resourceloader.

I'm adding this now as I'd like to rule out or confirm whether the
our own CDN caching plays a rule in the corruptions reported at
T321394. If so, we can dig deeper there. If not, then we're likely
looking at a problem relating to browser extensions, Wikipedia mirrors,
or user scripts; rather than something under our own control.

Bug: T321394
Change-Id: Iad2f5189da33551b59653c2a6783419d6ad955d0