preview: Maintain sort order of templates list in previews
When building the list of templates used in a preview, sometimes
extra system messages have to be fetched. If this is done, the
requests for these can take varying amounts of time (and be
redundant because often templates need the same messages), so
this changes the list-processing to be sequential.

Bug: T334441
Bug: T335624
Change-Id: I315345f9754cf50d65c159c5f73e20d4c5364b15