ResourceLoader: Micro-optim mw.loader `@import` support
Also improve the unit test which passed even if one were to
remove the `@import` check, because the first stylesheet
always works.

The test was originally well-written I believe, since we used to
actually re-use a single `<style>` tag for all modules, whereas
nowadays we batch it per animation frame and then never touch it
again because browsers incur significant recalculations if a style
tag string is mutated (requires re-parsing and re-calculating styles).

Change-Id: If42eb5d4f8011f07d43330e9c260413c43b28fce